fix: s3 delete 방어 로직

+++ src/main/java/com/haruse3kki/haruse3kki/service/S3Service.java
@@ -42,6 +42,11 @@ public class S3Service {
     }
 
     public void delete(String fileUrl) {
+
+        if (fileUrl == null || !fileUrl.contains(".amazonaws.com/")) {
+            return;
+        }
+
         String key = fileUrl.substring(fileUrl.indexOf(".amazonaws.com/") + ".amazonaws.com/".length());
 
         s3Client.deleteObject(
